+++ homematicip/group.py
@@ -276,6 +276,15 @@ class AlarmSwitchingGroup(Group):
         data = {"groupId": self.id, "signalOptical": signalOptical}
         return self._restCall("group/switching/alarm/setSignalOptical", body=json.dumps(data))
 
+    def test_signal_acoustic(self,
+                            signalAcoustic=AcousticAlarmSignal.FREQUENCY_FALLING):
+        data = {"groupId": self.id, "signalAcoustic": signalAcoustic}
+        return self._restCall("group/switching/alarm/testSignalAcoustic", body=json.dumps(data))
+
+    def set_signal_acoustic(self,
+                           signalAcoustic=AcousticAlarmSignal.FREQUENCY_FALLING):
+        data = {"groupId": self.id, "signalAcoustic": signalAcoustic}
+        return self._restCall("group/switching/alarm/setSignalAcoustic", body=json.dumps(data))
 
 # at the moment it doesn't look like this class has any special properties/functions
 # keep it as a placeholder in the meantime